refactor config, split to playbook and command
Also fix missing "tag" type in targets
