+++ app/config/playbook.go
@@ -1,7 +1,6 @@
 package config
 
 import (
-	"bytes"
 	"fmt"
 	"io"
 	"log"
@@ -20,7 +19,7 @@ import (
 	"github.com/umputun/simplotask/app/config/deepcopy"
 )
 
-// PlayBook defines top-level config yaml
+// PlayBook defines top-level config object
 type PlayBook struct {
 	User      string            `yaml:"user" toml:"user"`           // ssh user
 	SSHKey    string            `yaml:"ssh_key" toml:"ssh_key"`     // ssh key
@@ -48,6 +47,7 @@ type Target struct {
 	Hosts  []Destination `yaml:"hosts" toml:"hosts"`   // direct list of hosts to run commands on, no need to use inventory
 	Groups []string      `yaml:"groups" toml:"groups"` // list of groups to run commands on, matches to inventory
 	Names  []string      `yaml:"names" toml:"names"`   // list of host names to run commands on, matches to inventory
+	Tags   []string      `yaml:"tags" toml:"tags"`     // list of tags to run commands on, matches to inventory
 }
 
 // Task defines multiple commands runs together
@@ -59,22 +59,6 @@ type Task struct {
 	OnError  string `yaml:"on_error" toml:"on_error"`
 }
 
-// Cmd defines a single command
-type Cmd struct {
-	Name        string            `yaml:"name" toml:"name"`
-	Copy        CopyInternal      `yaml:"copy" toml:"copy"`
-	Sync        SyncInternal      `yaml:"sync" toml:"sync"`
-	Delete      DeleteInternal    `yaml:"delete" toml:"delete"`
-	Wait        WaitInternal      `yaml:"wait" toml:"wait"`
-	Script      string            `yaml:"script" toml:"script,multiline"`
-	Environment map[string]string `yaml:"env" toml:"env"`
-	Options     struct {
-		IgnoreErrors bool `yaml:"ignore_errors" toml:"ignore_errors"`
-		NoAuto       bool `yaml:"no_auto" toml:"no_auto"`
-		Local        bool `yaml:"local" toml:"local"`
-	} `yaml:"options" toml:"options,omitempty"`
-}
-
 // Destination defines destination info
 type Destination struct {
 	Name string   `yaml:"name" toml:"name"`
@@ -82,33 +66,6 @@ type Destination struct {
 	Port int      `yaml:"port" toml:"port"`
 	User string   `yaml:"user" toml:"user"`
 	Tags []string `yaml:"tags" toml:"tags"`
-}
-
-// CopyInternal defines copy command, implemented internally
-type CopyInternal struct {
-	Source string `yaml:"src" toml:"src"`
-	Dest   string `yaml:"dst" toml:"dst"`
-	Mkdir  bool   `yaml:"mkdir" toml:"mkdir"`
-}
-
-// SyncInternal defines sync command (recursive copy), implemented internally
-type SyncInternal struct {
-	Source string `yaml:"src" toml:"src"`
-	Dest   string `yaml:"dst" toml:"dst"`
-	Delete bool   `yaml:"delete" toml:"delete"`
-}
-
-// DeleteInternal defines delete command, implemented internally
-type DeleteInternal struct {
-	Location  string `yaml:"path" toml:"path"`
-	Recursive bool   `yaml:"recur" toml:"recur"`
-}
-
-// WaitInternal defines wait command, implemented internally
-type WaitInternal struct {
-	Timeout       time.Duration `yaml:"timeout" toml:"timeout"`
-	CheckDuration time.Duration `yaml:"interval" toml:"interval"`
-	Command       string        `yaml:"cmd" toml:"cmd,multiline"`
 }
 
 // Overrides defines override for task passed from cli
@@ -353,8 +310,8 @@ func (p *PlayBook) TargetHosts(name string) ([]Destination, error) {
 func (p *PlayBook) targetHosts(name string) ([]Destination, error) {
 	t, ok := p.Targets[name] // get target from playbook
 	if ok {
-		if len(t.Hosts) == 0 && len(t.Names) == 0 && len(t.Groups) == 0 {
-			return nil, fmt.Errorf("target %q has no hosts, names or groups", name)
+		if len(t.Hosts) == 0 && len(t.Names) == 0 && len(t.Groups) == 0 && len(t.Tags) == 0 {
+			return nil, fmt.Errorf("target %q has no hosts, names, tags or groups", name)
 		}
 		// we have found target in playbook, process hosts, names and group
 		res := []Destination{}
@@ -381,6 +338,22 @@ func (p *PlayBook) targetHosts(name string) ([]Destination, error) {
 			for _, g := range t.Groups {
 				// we don't set default port and user here, as they are set in inventory already
 				res = append(res, p.inventory.Groups[g]...)
+			}
+		}
+
+		if len(t.Tags) > 0 && p.inventory != nil {
+			// target has "tags", get all hosts from inventory for each tag
+			for _, tag := range t.Tags {
+				for _, h := range p.inventory.Groups[allHostsGrp] {
+					if len(h.Tags) == 0 {
+						continue
+					}
+					for _, t := range h.Tags {
+						if strings.EqualFold(t, tag) {
+							res = append(res, h)
+						}
+					}
+				}
 			}
 		}
 
@@ -532,93 +505,6 @@ func (p *PlayBook) loadInventory(loc string) (*InventoryData, error) {
 	}
 
 	return &data, nil
-}
-
-// GetScript returns a script string and an io.Reader based on the command being single line or multiline.
-func (cmd *Cmd) GetScript() (string, io.Reader) {
-	if cmd.Script == "" {
-		return "", nil
-	}
-
-	elems := strings.Split(cmd.Script, "\n")
-	if len(elems) > 1 {
-		return "", cmd.getScriptFile()
-	}
-
-	return cmd.getScriptCommand(), nil
-}
-
-// GetScriptCommand concatenates all script line in commands into one a string to be executed by shell.
-// Empty string is returned if no script is defined.
-func (cmd *Cmd) getScriptCommand() string {
-	if cmd.Script == "" {
-		return ""
-	}
-
-	envs := cmd.genEnv()
-	res := "sh -c \""
-	if len(envs) > 0 {
-		res += strings.Join(envs, " ") + " "
-	}
-
-	elems := strings.Split(cmd.Script, "\n")
-	var parts []string // nolint
-	for _, el := range elems {
-		c := strings.TrimSpace(el)
-		if len(c) < 2 {
-			continue
-		}
-		if i := strings.Index(c, "#"); i > 0 {
-			c = strings.TrimSpace(c[:i])
-		}
-		parts = append(parts, c)
-	}
-	res += strings.Join(parts, "; ") + "\""
-	return res
-}
-
-// GetScriptFile returns a reader for script file. All the line in the command used as a script, with hashbang,
-// set -e and environment variables.
-func (cmd *Cmd) getScriptFile() io.Reader {
-	var buf bytes.Buffer
-
-	buf.WriteString("#!/bin/sh\n") // add hashbang
-	buf.WriteString("set -e\n")    // add 'set -e' to make the script exit on error
-
-	envs := cmd.genEnv()
-	// set environment variables for the script
-	if len(envs) > 0 {
-		for _, env := range envs {
-			buf.WriteString(fmt.Sprintf("export %s\n", env))
-		}
-	}
-
-	elems := strings.Split(cmd.Script, "\n")
-	for _, el := range elems {
-		c := strings.TrimSpace(el)
-		if len(c) < 2 {
-			continue
-		}
-		if strings.HasPrefix(c, "#") {
-			continue
-		}
-		if i := strings.Index(c, "#"); i > 0 {
-			c = strings.TrimSpace(c[:i])
-		}
-		buf.WriteString(c)
-		buf.WriteString("\n")
-	}
-
-	return &buf
-}
-
-func (cmd *Cmd) genEnv() []string {
-	envs := make([]string, 0, len(cmd.Environment))
-	for k, v := range cmd.Environment {
-		envs = append(envs, fmt.Sprintf("%s='%s'", k, v))
-	}
-	sort.Slice(envs, func(i, j int) bool { return envs[i] < envs[j] })
-	return envs
 }
 
 // checkConfig checks validity of config
